# apps/api/openapi.yaml
openapi: 3.0.0
paths:
...
  /api/v1/patients/{id}:
    get:
      operationId: PatientsController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      tags:
        - Patients
      x-required-permissions:
        - patient.view
    patch:
      operationId: PatientsController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePatientDto'
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      tags:
        - Patients
      x-required-permissions:
        - patient.edit
...
  /api/v1/shifts:
    get:
      operationId: ShiftsController_findAll
      parameters: []
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      tags:
        - Scheduling - Shifts
      x-any-permissions:
        - scheduling.manage
        - nursing.ward.view
  /api/v1/shifts/{id}/capacity:
    patch:
      operationId: ShiftsController_updateCapacity
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateShiftCapacityDto'
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      tags:
        - Scheduling - Shifts
      x-required-permissions:
        - shift.manage
...
  /api/v1/patients/{patientId}/supply-profile:
    get:
      operationId: PatientSupplyProfileController_findAll
      parameters:
        - name: patientId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      tags:
        - Inventory - Patient Supply Profile
      x-required-permissions:
        - inventory.view
    put:
      operationId: PatientSupplyProfileController_setProfile
      parameters:
        - name: patientId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetSupplyProfileDto'
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      tags:
        - Inventory - Patient Supply Profile
      x-required-permissions:
        - inventory.manage
  /api/v1/patients/{patientId}/supply-profile/{itemId}:
    delete:
      operationId: PatientSupplyProfileController_removeItem
      parameters:
        - name: patientId
          required: true
          in: path
          schema:
            type: string
        - name: itemId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      tags:
        - Inventory - Patient Supply Profile
      x-required-permissions:
        - inventory.manage
...
  /api/v1/machines/{id}:
    get:
      operationId: MachinesController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      tags:
        - Machines
      x-required-permissions:
        - machine.view
...
  /api/v1/nursing/assignments:
    post:
      operationId: AssignmentsController_upsert
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAssignmentDto'
      responses:
        '201':
          description: ''
      security:
        - bearer: []
      tags:
        - Nursing - Assignments
      x-required-permissions:
        - nursing.assign
    get:
      operationId: AssignmentsController_listForWard
      parameters:
        - name: wardId
          required: true
          in: query
          schema:
            type: string
        - name: shiftId
          required: false
          in: query
          schema:
            nullable: true
            type: string
        - name: date
          required: true
          in: query
          schema:
            type: string
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      tags:
        - Nursing - Assignments
      x-any-permissions:
        - nursing.assign
        - nursing.ward.view.all
  /api/v1/nursing/nurses:
    get:
      operationId: AssignmentsController_listNurseCandidates
      parameters: []
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      tags:
        - Nursing - Assignments
      x-required-permissions:
        - nursing.assign
  /api/v1/nursing/my-assignments:
    get:
      operationId: AssignmentsController_listMine
      parameters:
        - name: date
          required: true
          in: query
          schema:
            type: string
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      tags:
        - Nursing - Assignments
...
  /api/v1/lab/tests/{testId}/trend:
    get:
      operationId: LabTrendController_trend
      parameters:
        - name: testId
          required: true
          in: path
          schema:
            type: string
        - name: patientId
          required: true
          in: query
          schema:
            type: string
        - name: limit
          required: false
          in: query
          schema:
            minimum: 1
            maximum: 100
            nullable: true
            type: integer
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      tags:
        - Lab - Trends
      x-required-permissions:
        - patient.view
...
  /api/v1/machines/{machineId}/timeline:
    get:
      operationId: MachineTimelineController_timeline
      parameters:
        - name: machineId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      tags:
        - Maintenance - Machine Timeline
      x-required-permissions:
        - machine.view
  /api/v1/machines/{machineId}/downtime:
    get:
      operationId: MachineTimelineController_downtime
      parameters:
        - name: machineId
          required: true
          in: path
          schema:
            type: string
        - name: from
          required: true
          in: query
          schema:
            type: string
        - name: to
          required: true
          in: query
          schema:
            type: string
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      tags:
        - Maintenance - Machine Timeline
      x-required-permissions:
        - machine.view
...
  /api/v1/reports/machines/{machineId}/downtime:
    get:
      operationId: MachineReportsController_downtime
      parameters:
        - name: machineId
          required: true
          in: path
          schema:
            type: string
        - name: format
          required: false
          in: query
          schema:
            nullable: true
            type: string
            enum:
              - pdf
              - excel
        - name: from
          required: false
          in: query
          schema:
            pattern: ^\d{4}-\d{2}-\d{2}$
            nullable: true
            type: string
        - name: to
          required: false
          in: query
          schema:
            pattern: ^\d{4}-\d{2}-\d{2}$
            nullable: true
            type: string
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      tags:
        - Reports - Machines
      x-any-permissions:
        - machine.view
        - maintenance.manage
...
  /api/v1/reports/operations/stock-movements:
    get:
      operationId: OperationalReportsController_stockMovements
      parameters:
        - name: format
          required: false
          in: query
          schema:
            nullable: true
            type: string
            enum:
              - pdf
              - excel
        - name: from
          required: false
          in: query
          schema:
            pattern: ^\d{4}-\d{2}-\d{2}$
            nullable: true
            type: string
        - name: to
          required: false
          in: query
          schema:
            pattern: ^\d{4}-\d{2}-\d{2}$
            nullable: true
            type: string
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      tags:
        - Reports - Operations
      x-required-permissions:
        - inventory.view
  /api/v1/reports/operations/drug-consumption:
    get:
      operationId: OperationalReportsController_drugConsumption
      parameters:
        - name: format
          required: false
          in: query
          schema:
            nullable: true
            type: string
            enum:
              - pdf
              - excel
        - name: month
          required: true
          in: query
          schema:
            pattern: ^\d{4}-\d{2}$
            type: string
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      tags:
        - Reports - Operations
      x-required-permissions:
        - pharmacy.dispense
  /api/v1/reports/operations/lab-volume:
    get:
      operationId: OperationalReportsController_labVolume
      parameters:
        - name: format
          required: false
          in: query
          schema:
            nullable: true
            type: string
            enum:
              - pdf
              - excel
        - name: from
          required: false
          in: query
          schema:
            pattern: ^\d{4}-\d{2}-\d{2}$
            nullable: true
            type: string
        - name: to
          required: false
          in: query
          schema:
            pattern: ^\d{4}-\d{2}-\d{2}$
            nullable: true
            type: string
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      tags:
        - Reports - Operations
      x-required-permissions:
        - lab.queue.view
...
info:
  title: DCMS API
  description: >-
    Dialysis Center Management System - REST API covering all 16 project phases
    (docs/PROJECT-PHASES-PLAN.md). Every protected route requires the bearer JWT
    returned by POST /auth/login, and every request is re-checked against the
    caller's current permissions - nothing here is cached from login time.
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    UpdatePatientDto:
      type: object
      properties:
        fullName:
          type: string
        gender:
          type: string
          enum:
            - MALE
            - FEMALE
        dateOfBirth:
          type: string
        phone:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        fileNumber:
          type: string
          nullable: true
        status:
          type: string
          enum:
            - ACTIVE
            - INACTIVE
            - DECEASED
            - TRANSFERRED
        dialysisStartDate:
          type: string
          nullable: true
        dryWeight:
          type: number
          nullable: true
          minimum: 1
          maximum: 300
        vascularAccessType:
          type: string
          enum:
            - FISTULA
            - CATHETER
            - GRAFT
          nullable: true
        vascularAccessLocation:
          type: string
          nullable: true
        diagnoses:
          type: string
          nullable: true
        chronicDiseases:
          nullable: true
          type: array
          items:
            type: string
        allergies:
          type: string
          nullable: true
        medicalNotes:
          type: string
          nullable: true
        specialInstructions:
          type: string
          nullable: true
        reason:
          type: string
          nullable: true
...
    UpdateShiftCapacityDto:
      type: object
      properties:
        nominalCapacity:
          type: integer
          minimum: 0
        reservedCapacity:
          type: integer
          nullable: true
          minimum: 0
        lateThresholdMinutes:
          type: integer
          nullable: true
          minimum: 0
      required:
        - nominalCapacity
...
    SupplyProfileEntryDto:
      type: object
      properties:
        itemId:
          type: string
        defaultQuantity:
          type: number
          minimum: 0.01
      required:
        - itemId
        - defaultQuantity
    SetSupplyProfileDto:
      type: object
      properties:
        entries:
          maxItems: 50
          type: array
          items:
            $ref: '#/components/schemas/SupplyProfileEntryDto'
      required:
        - entries
...
    CreateAssignmentDto:
      type: object
      properties:
        wardId:
          type: string
        shiftId:
          type: string
        date:
          type: string
        nurseId:
          type: string
        patientIds:
          type: array
          items:
            type: string
      required:
        - wardId
        - shiftId
        - date
        - nurseId
        - patientIds
...